Add frontend scene list changed listener
If a scene is added this will just create a new entry in the
scene_matrix map.

If a scene is removed this will check for any transition override that
contains that scene and it will be removed.

If a scene is edited this will update all map entrys that contain the
old name to the new scene name.